kleaf: Support given empty srcs in kernel_module_group
In some build conditions, BUILD.bazel might call kernel_module_group
with empty srcs. This is because the kernel module group needs to be
able to handle cases where there are no source files for a particular
module. In these cases, the kernel_module_info object for the
module should be set to KernelModuleInfo(). This will cause the kernel
module group to skip the module when building the kernel.

For other members of the kernel module group, they will be fine if
the targets are empty. This is because the kernel module group will
only build the modules that have source files.

Test: Build Pass and Boot to Android w/ and w/o bazel package in srcs
Bug: 275494044
Change-Id: I2e918cf7a40cf9c48ec4c14978e3f8bbdada53fe
Signed-off-by: Jerry Pai <jerrypai@google.com>