kleaf: Add abi_definition_xml arg to kernel_abi
Currently, if a kernel tree contains an ABI definition in XML format,
the Bazel analysis fails. This is because when abi_definition_xml is
set and passed down to other rules, kernel_abi does not have an arg
that can handle it.

Add abi_definition_xml to kernel_abi to fix this issue.

Bug: 275697885
Change-Id: I5704fc0001c82e08717d7ad4444da8ab7ea3ab87
Signed-off-by: John Moon <quic_johmoo@quicinc.com>