kleaf: Fix run kernel_config when defconfig is generated.
When the defconfig file is generated, the
arch/$SRCARCH/configs/$DEFCONFIG file does not
exist beforehand. Hence, there is no need to
record its symlink and restore it later.

In this case, simply print the path of
the new defconfig file if it exists.

If it doesn't exist, it may be the case that POST_DEFCONFIG_CMDS
already moved it away.

See microdroid for an example on how to make this properly
work with generated defconfig files.

Bug: 275468906
Bug: 276889737

Change-Id: If0b94f577ad064c4ef06329102033717c3e8c899