kleaf: Use STABLE_SOURCE_DATE_EPOCHS
This change turns STABLE_SOURCE_DATE_EPOCH into
STABLE_SOURCE_DATE_EPOCHS, which is a mapping that
is keyed on the path.

This change prepares for a follow up change that uses
repo list to get the list of projects.

In addition, this change also fixes a bug that
SOURCE_DATE_EPOCH in the environment was previously
not considered.

Bug: 277367919
Test: manual on Pixel

Change-Id: Id16041ce23821539016df7584c478dc18aac4fa0