kleaf: Handles subdirectories for scmversion / source_date_epoch.
When KERNEL_DIR or an external module is defined in
a subdirectory of the git repository, we need to
check parent directories in the git metadata map.

For example, if we have git project private/tuna/nfc
and kernel_module is defined in private/tuna/nfc/mod,
then STABLE_SCMVERSIONS contains:

  private/tuna/nfc:-g123456abcdef

However, for the kernel_module, ext_mod=private/tuna/nfc/mod.
Hence, we need to check if private/tuna/nfc/mod is in the map,
then private/tuna/nfc, and so on.

Test: TH
Bug: 277367919
Change-Id: I98fdab87ad79a62a690331795819f2146016453b