kleaf: split KernelEnvInfo.dependencies.
The kernel_env now exposes KernelEnvInfo
with dependencies split into two depsets: inputs and tools.

The kernel_module now exposes KernelModuleSetupInfo with
fields inputs (which indicate that this is built for the
target platform) and a setup fragment cmd.

This is so that we correctly categorizes files for the
target and for the execution platform.

Bug: 263385781
Test: TH

Change-Id: I379397ee78bfa2a2b7495ed2ccd4b7e324f26f51