kleaf: fix case when workspace root is not the same as repo root.
In the rare case when the workspace root is a subdir
of the repo root, workspace_status_stamp.py should handle paths
properly.

This does not affect --repo_manifest option. If the repo root
is not the same as the workspace root, AND repo is not available
during build time, the repo manifest must be preprocessed so
the paths are fixed up properly.

Fixes: 281675905
Bug: 281675905

Test: TH
Change-Id: Id1276d69b9a9e7cf159b61e3deb9b3494b36dcbb