kleaf: DDK: put cflags in file
When there are too many cflags, the build may fail with
the following error:

  make[3]: /bin/sh: Argument list too long

This is usually caused by the super long include paths
in the clang command line, though it is possible to be
caused by others. Hence, we put some of the generated flags
into a .cflags file, then provide it to clang with the @
syntax.

Caveats:
- Environment variables are not allowed in .cflags files,
  so we need to replace references to $(ROOT_DIR) manually.

Bug: 282786692
Change-Id: I1df4cc928970b5dd3e6a104c7a136957b03ba378