kleaf: mixed kernel_build module_staging_archive handles scmversion mismatch.
When the base_kernel is built with a different scmversion than
the device's kernel_build, the module_staging_archive for the
base_kernel puts GKI modules in a different directory. Hence,
the original code to exclude non-overriden GKI modules from the
module staging archive no longer works.

Rework the code so we run `basename` on the list of GKI modules. Then,
extract the following:

- selected (non-overriden) GKI modules from base_kernel's staging dir.
  - The GKI modules are moved to the /lib/modules/<device's UTS release>
    directory; this may mismatch with `uname -r`
- modules from the device build's staging dir, minus the selected
  (non-overriden) GKI modules

Caveat: After this patch, for mixed builds, modules are always
installed to /lib/modules/<device's UTS release>, where the
device's UTS release is defined by KERNELVERSION when building
in-tree modules and device tree files (that is, when building
the device's kernel_build target). This may not be the same as
the `uname -r` from vmlinux.

If this change breaks a boot process for builds where vmlinux
and the device's kernel_build are built from different trees,
it may be because that init
(insmod) loads modules from `/lib/modules/$(uname -r)`. It should be
updated to use `/lib/modules/*` instead.

Test: manual by building a Pixel device against prebuilts,
      which are built on ci.android.com with a different scmversion.
Bug: 284198523

Change-Id: I49a15247a85109328093df8701378b00b6d8bf35