kleaf: mixed kernel_build module_staging_archive handles scmversion mismatch.
When the base_kernel is built with a different scmversion than
the device's kernel_build, the module_staging_archive for the
base_kernel puts GKI modules in a different directory. Hence,
the original code to exclude non-overriden GKI modules from the
module staging archive no longer works.

Rework the code so we run `basename` on the list of GKI modules. Then,
extract the following:

- modules from the device build's staging dir, minus the selected
  (non-overriden) GKI modules
- selected (non-overriden) GKI modules from base_kernel's staging dir.

Test: manual by building a Pixel device against prebuilts,
      which are built on ci.android.com with a different scmversion.
Bug: 284198523

Change-Id: I49a15247a85109328093df8701378b00b6d8bf35