kleaf: write_depset provides a depset of all necessary dependencies.
To use the content of write_depset, the original depset
should also be part of the dependencies of the action. Previously,
for ddk_config, the original depset is added elsewhere. Make it
clear by having write_depset also provides a depset that
contains the original depset and the written file.

Test: TH
Bug: 286589887
Change-Id: I5cb0d280442b5a9c8cbfed73d5b304bd83bea211