kleaf: Add --defconfig_fragment flag
Add a --defconfig_fragment flag that adds to all
kernel_build.defconfig_fragments.

The result is additive, not overriding, because the
corresponding attribute is a list. However, the flag
itself only accepts 1 label, not multiple (because there's no
allow_multiple label_flag). For multiple files, the
device maintainer is expected to define a separate filegroup,
which could in turn be controlled by additional flags
(i.e. a filegroup() with select()'ed sources).

Bug: 286589887
Test: manual
Change-Id: I061d36e0a701d7f804340a734df7e88928429802