Revert "Revert "kleaf: Set --incompatible_kernel_use_resolved_toolchains by default.""
This reverts commit 6794aaa7ca8d853568bc079b35e20ddb44d9c40a.

Reason for revert: Relanding change

Below is the original commit message.

This change ensures that we use cflags / ldflags from
the resolved toolchains, not _setup_env.sh.

This requires kernel_build.arch to be set properly.

This behavior can be reverted by
--incompatible_kernel_use_resolved_toolchains
temporarily. If one needs this flag, they must report
a bug.

Change-Id: I20756fdce6d8f58fe6a54acac1b7b28da65f0c09
Test: TH
Bug: 272164611