kleaf: Add --debug_cache_dir_conflict.
Useful flag to debug issues with --cache_dir conflicts.

If a user hits a flaky issue that seem to be related to cache
dir conflicts, they may use the following:

--debug_cache_dir_conflict=detect to fail early if there
is indeed a conflict.

--debug_cache_dir_conflict=resolve to serialize the build
(attempt to resolve the error) if there is any conflict.

This flag requires `flock` to exist on the host machine.
Bug 289137454 tracks changing flock's implementation to the toybox one.

Test: manually build with --debug_cache_dir_conflict=detect
Test: lock the file, manually build with
  --debug_cache_dir_conflict=resolve, build hangs; then unlock,
  build proceeds successfully.
Bug: 266956564
Bug: 289137454
Change-Id: I945edf130c0811b25634aa8feb2b85eb726c25ab