kleaf: Support kernel_build.page_size
If kernel_build sets a page_size attribute,
the value will be used as a fallback if --page_size
is not set.

The logic is as follows:
- If --page_size is set, use its value.
- If --page_size is not set (or set to default), use
  value of kernel_build.page_size

Implementation detail: technically,
kernel_build.page_size is configurable as long
as the value of the dictionary to `select()` is
not `None`, but "4k". However, we hardly need this
attribute to be configurable, so mark as non-configurable
for now.

Bug: 286589887
Change-Id: Icc2f2d14444727ead4d2c937c6e4423feea0ac5b