kleaf: hermetic_tools: Do not use old fields.
Stop using the old fields in hermetic_tools:
srcs, py3_outs, host_tools, tar_args, rsync_args.

This removes most of the dependency on the symlinks under
build/kernel/build-tools/path. Instead of using
the symlinks, point to the actual binary directly.

The symlinks build/kernel/build-tools/path are pending deletion.
However, there are still users of the symlinks, so we
are keeping the symlinks for now.

After this change, most of the symlinks under
build/kernel/build-tools/path/linux-x86
won't show up in the sandbox when we build the kernel.

The list is created by doing the following:

  for link in $(find build-tools/path/linux-x86/ -type f -o -type l); do
    if [[ -L $link ]]; then
      real=$(readlink $link)
      if [[ $(basename $real) != $(basename $link) ]]; then
        echo $real: $(basename $link)
      else
        echo $real
      fi
    else
      echo $link
    fi
  done | sort

... then manually replacing the result with proper labels.

In addition, the following symlinks originally points to toybox binary
directly. Now, to reduce dependency on build-tools/path, we use
the symlinks under prebuilts/build-tools instead:
  cpio, realpath, tr

In addition, dtc is not resolved because build.config.common
hardcoded DTC to the symlink.

Test: TH
Bug: 291816237
Change-Id: Icd7ff7678fe9d136772c9ac75ac5c6190fda07d3