kernel_sbom: add support to pass a list of files
Add support for supplying a list of files --files <file1> <file2> ... as
an alternative to specifying a distribution directory. This is useful to
avoid all sorts of issues with globbing in a forest of symbolic links as
is the case for bazel based builds. Bazel can precisely supply all
inputs that will get distributed though. Those become mutually exclusive
options and at least one is required.

Bug: 279663374
Change-Id: I883e6d12535eea6fc7d71f8ce9173e55fac73684
Signed-off-by: Matthias Maennich <maennich@google.com>