kleaf: hermetic_tools: Add original tools in symlinks to sandbox.
This change is needed because build.config.common hardcodes
path to DTC instead of respecting DTC in PATH. Hence, the "source"
of the following key in symlinks:

  "build-tools/path/linux-x86/dtc": "dtc"

... also needs to be added.

The symlinks under build-tools/path/linux-x86/ are added
for backwards compatibility reasons. Once we confirm that
the symlinks are no longer needed, the flag
--incompatible_disable_hermetic_tools_symlink_source should
be enabled by default.

Bug: 291816237
Change-Id: Id371a6aca8cfb06417f5d780d9a6ae0998052de7