kleaf: Implement checkpatch().
This macro can be placed in the root of any Git
directory in the source tree to enable checkpatch.

For example, if the target is placed in //common,
one can

  bazel run //common:checkpatch

to check HEAD.

Additional flags like --bid or --dist_dir may be provided
in presubmit.

Bug: 259995152
Change-Id: If58d9e15fff53f8349663525c710d6cb87343c2e
(cherry picked from commit 93c6b8f73e189952ccbff8778639d705f3486842)