Support not running checkpatch on some projects.
The checkpatch script for android-common-mainline runs checkpatch
on all the git projects that contain commits. But there are some
projects that it doesn't make sense to run checkpatch on.
One such project is kernel/tests. It doesn't use kernel coding
conventions, and also it doesn't have a BUILD file so the build
breaks with:

ERROR: no such package 'kernel/tests': BUILD file not found in any of the following directories. Add a BUILD file to a directory to mark it as a package.

Test: Put an applied.prop containing a SHA1 for kernel/tests in /tmp/dist/applied
Test: tools/bazel run --verbose_failures --jobs=80 --make_jobs=80 --kleaf_localversion --config=android_ci //build/kernel/static_analysis:checkpatch_presubmit -- --dist_dir /tmp/dist
Change-Id: Id98ea6a7e3c1e84d914dd23bd10479859172cf9a