kleaf: Skip checkpatch_presubmit if project has no BUILD file.
The checkpatch script for android-common-mainline runs checkpatch
on all the git projects that contain commits. But there are some
projects that it doesn't make sense to run checkpatch on.
One such project is kernel/tests. It doesn't use kernel coding
conventions, and also it doesn't have a BUILD file so the build
breaks with:

ERROR: no such package 'kernel/tests': BUILD file not found in any of the following directories. Add a BUILD file to a directory to mark it as a package.

Test: Put an applied.prop containing a SHA1 for kernel/tests in /tmp/dist/applied
Bug: 259995152
Change-Id: I1f9d0cffb0ed0a6fc47830deb3d838108d8c317f