Fixes for python 3.11
Bazel previously had a bug where the source directory of the entrypoint
python script would be added to the path. They added PYTHONSAFEPATH
to the enviornment to fix it, which was introduced in python 3.11.

Now that python is being updated to 3.11, the fix is in, which breaks
some scripts that don't properly specify import folders.

Add the import folders where required.

Bug: 278602456
Test: tools/bazel run --config=android_ci //common:kernel_aarch64_abi_dist -- --flat
Test: bazel test build/kernel/kleaf:check_declared_output_list_test
Test: bazel test build/kernel/kleaf/impl:get_kmi_string_test
Change-Id: I6c5c83c1eaafcb5dccb6ac6ee9716e1656b2d652
(cherry picked from commit 6b09ffdec0419b3776a4592c614e84f58d8c50f0)