kleaf: kernel_modules_install has less dep on modules_staging_archive
This is so that kernel_modules_install does not
use modules_staging_archive from kernel_build
besides extraction. That is, for check_duplicated_files_in_archives,
we no longer require the path to modules_staging_archive.

This is so that the extraction of modules_staging_archive
can later be moved into KernelBuildExtModuleInfo.

Bug: 291918087
Change-Id: I8e7eee4547e62d038b533cd31de096c097cd7ccf