kleaf: bazel.py determines --output_root against abs workspace.
Add a new variable, workspace_dir, that refers to
the top-level workspace. This is different from
kleaf_repo_dir when Kleaf is in a sub-repository.

If Kleaf tools are in a sub-repository (external repository),
the directory structure becomes the following:

- /<workspace_dir>
    - WORKSPACE
    - out/bazel/output_user_root/<sha>/external/kleaf (kleaf_repo_dir)
        - build/kernel
	- ...

This change ensures we correctly locate some directories against
the workspace root. In particular, --output_root is by default
<workspace>/out. This has the following effects:
- The cache dir, by default, lives in <workspace>/out/cache.
- output_user_root, by default, lives in
  <workspace>/out/bazel/output_user_root
- javatmp, by default, lives in <workspace>/out/bazel/javatmp.

Performance analysis:

On average, this change adds 0.03s to every bazel command.
  `time bazel info workspace` with bazel server started:
      increases from 0.123s to 0.155s (average over 100 experiments).
  `bazel shutdown; time bazel info workspace` (cold start):
      the second command increases from 1.435s to 1.461s
      (average over 100 experiments).

Bug: 293260835
Test: TH
Change-Id: I29cdfb50f86047268e5157b047ee54eafab2d872