kleaf: Add support for kernel_build.sanitizers
Currently, the only way to configure a sanitizer (e.g. kasan, kcsan,
etc.) is to pass the corresponding flag. In some cases, it would be
better to specify the sanitizer as part of the kernel_build rule.

Add an attribute to kernel_build for sanitizers. Currently, this
attribute can only accept one value, but is implemented as a list
for future expansion.

With this solution, the unit tests that check analysis phase
errors are no longer correct, so they are disabled.

Bug: 302126333
Change-Id: I60b46a40287b57a1a24f2e8823a8cd4382402d7d
Signed-off-by: John Moon <quic_johmoo@quicinc.com>
Signed-off-by: Yifan Hong <elsk@google.com>
(cherry picked from commit f4696eef1f3fccd45dcc77595344fcc1fe5bc337)