kleaf: artifacts test do not depend on hermetic_tools aliases.
Such aliases have been deprecated. This ensures
that running these tests won't produce warning
messages because they depend on the deprecated
aliases.

To achieve this, I created a drop-in replacement
for py_test that sets the hermetic toolchain
in PATH before running the test. Then I modified
the tests to search for the binary in PATH.

Test: TH
Bug: 234125794
Change-Id: Id84321227e6af04e2433516c30f8b11cbb9bcafe