WIP: kleaf: script to check-in sub BCR
This is a helper script to check in a sub
tree of the Bazel Central Regsitry (BCR).

After this is checked-in to the tree, one
can use --registry=file:///path/to/registry,
so they don't have to download the registry
from the internet.

This serves as a workaround before local vendoring
is supported for bzlmod.

Test: manual. Run the following:
  build/kernel/kleaf/registry/update.py
  bazel build //common:kernel_aarch64 --enable_bzlmod --registry=file://$(bazel info workspace)/build/kernel/kleaf/registry --config=internet

Change-Id: I7a1cfcb056e3c7f95c18ff3c0feb085bd52d7013