rust: Enable Sample modules
Until Rust binder lands, these will act as a canary for whether Rust
support is working correctly. Without this, we only know that the
support builds, not that it can actually run.

We can revert this patch when Binder is enabled.

Bug: 302556490
Change-Id: I8664355416dcde3281b93b6c617a23b5f7842ab4
Merged-In: I8664355416dcde3281b93b6c617a23b5f7842ab4