Kleaf: Build parts of the prebuilt toolchain from sources
Add a flag `--toolchain_from_sources` that guards the source-built
prebuilts for now, but when the flag is enabled, built parts of the
otherwise prebuilt toolchain from sources in order to eventually get rid
of the prebuilt tools altogether.

Begin with an empty list, rendering this change effectively a safe noop.

Change-Id: Ibdd6bfebf0dfb756367a79a03ca02e6d465a7298
Signed-off-by: Matthias Maennich <maennich@google.com>