Kleaf: Build parts (lz4) of the prebuilt toolchain from sources
Add a flag `--toolchain_from_sources` that guards the source-built
prebuilts for now, but when the flag is enabled, built parts of the
otherwise prebuilt toolchain from sources in order to eventually get rid
of the prebuilt tools altogether.

Begin with `lz4` that does not have any other dependencies and exposes a
fairly simple build.

Change-Id: Ibdd6bfebf0dfb756367a79a03ca02e6d465a7298
Signed-off-by: Matthias Maennich <maennich@google.com>