kleaf: common_kernels: Ignore cscope & tags
As the common_kernel_sources filegroup includes _any_ file in the tree,
with the exception of Git and Bazel files, the build system gets
confused if some files are present at the time where the glob is
evaluated but absent in following build commands. An example of this is
an IDE automatically updating the code index file when saving a source
file (e.g. tags.lock or tags.temp) as the kernel is getting built.

Given that cscope and tags files are officially supported by the Linux
Makefiles and as Bazel doesn't actually need them to build the kernel,
exclude them from the filegroup.

Test: build the kernel
Change-Id: I1235f074bf40c815826e9ca16a70c196f8b204b7