kleaf: Fix generation of modules.load.[recovery|charger]
Commit d08605f4c795 ("kleaf: Add support for specifying modules for
recovery/charger") accidentally made it so that
modules.load.[recovery|charger] are only generated if the vendor_boot
or vendor_kernel_boot images are made. Similar to modules.load, it is
possible to want to generate modules.load.[recovery|charger] without
a vendor_boot or vendor_kernel_boot image, so decouple the two.

Bug: 324356826
Fixes: d08605f4c795 ("kleaf: Add support for specifying modules for recovery/charger")
Change-Id: Idf9db5954e8c1bae01915f104c704302dbfd81ae
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>